Reorganize registration unit tests
