+++ registrations/tests/test_registration_put.py
@@ -2,8 +2,10 @@ import pytest
 from rest_framework import status
 
 from events.tests.utils import versioned_reverse as reverse
-from registrations.tests.test_registration_admin_side import get_event_url
-from registrations.tests.test_registration_post import create_registration
+from registrations.tests.test_registration_post import (
+    create_registration,
+    get_event_url,
+)
 
 # === util methods ===
