+++ registrations/tests/test_signup_get.py
@@ -4,6 +4,7 @@ from rest_framework import status
 from events.tests.conftest import APIClient
 from events.tests.utils import versioned_reverse as reverse
 from registrations.models import SignUp
+from registrations.tests.test_signup_post import assert_create_signup
 
 # === util methods ===
 
@@ -265,3 +266,107 @@ def test__signup_list_assert_text_filter(
         response.data["detail"]
         == "attendee_status can take following values: waitlisted, attending, not invalid-value"
     )
+
+
+@pytest.mark.django_db
+def test_filter_signups(
+    api_client, registration, registration2, user, user2, event, event2
+):
+    registration.audience_min_age = None
+    registration.audience_max_age = None
+    registration.save()
+    registration2.audience_min_age = None
+    registration2.audience_max_age = None
+    registration2.save()
+
+    api_client.force_authenticate(user=None)
+    sign_up_payload = {
+        "registration": registration.id,
+        "name": "Michael Jackson",
+        "email": "test@test.com",
+    }
+    sign_up_payload1 = {
+        "registration": registration.id,
+        "name": "Michael Jackson1",
+        "email": "test1@test.com",
+    }
+    sign_up_payload2 = {
+        "registration": registration.id,
+        "name": "Michael Jackson2",
+        "email": "test2@test.com",
+    }
+    sign_up_payload3 = {
+        "registration": registration.id,
+        "name": "Michael Jackson3",
+        "email": "test3@test.com",
+    }
+    sign_up_payload4 = {
+        "registration": registration2.id,
+        "name": "Joe Biden",
+        "email": "test@test.com",
+        "extra_info": "cdef",
+    }
+    sign_up_payload5 = {
+        "registration": registration2.id,
+        "name": "Hillary Clinton",
+        "email": "test1@test.com",
+        "extra_info": "abcd",
+    }
+    sign_up_payload6 = {
+        "registration": registration2.id,
+        "name": "Donald Duck",
+        "email": "test2@test.com",
+        "membership_number": "1234",
+    }
+    sign_up_payload7 = {
+        "registration": registration2.id,
+        "name": "Mickey Mouse",
+        "email": "test3@test.com",
+        "membership_number": "3456",
+    }
+    response = assert_create_signup(api_client, registration.id, sign_up_payload)
+    signup = SignUp.objects.get(pk=response.data["attending"]["people"][0]["id"])
+    response = assert_create_signup(api_client, registration.id, sign_up_payload1)
+    signup1 = SignUp.objects.get(pk=response.data["attending"]["people"][0]["id"])
+    response = assert_create_signup(api_client, registration.id, sign_up_payload2)
+    signup2 = SignUp.objects.get(pk=response.data["attending"]["people"][0]["id"])
+    response = assert_create_signup(api_client, registration.id, sign_up_payload3)
+    signup3 = SignUp.objects.get(pk=response.data["attending"]["people"][0]["id"])
+    response = assert_create_signup(api_client, registration2.id, sign_up_payload4)
+    signup4 = SignUp.objects.get(pk=response.data["attending"]["people"][0]["id"])
+    response = assert_create_signup(api_client, registration2.id, sign_up_payload5)
+    signup5 = SignUp.objects.get(pk=response.data["attending"]["people"][0]["id"])
+    response = assert_create_signup(api_client, registration2.id, sign_up_payload6)
+    signup6 = SignUp.objects.get(pk=response.data["attending"]["people"][0]["id"])
+    response = assert_create_signup(api_client, registration2.id, sign_up_payload7)
+    signup7 = SignUp.objects.get(pk=response.data["attending"]["people"][0]["id"])
+
+    api_client.force_authenticate(user)
+    get_list_and_assert_signups(
+        api_client, registration.id, "", [signup, signup1, signup2, signup3]
+    )
+
+    api_client.force_authenticate(user2)
+    get_list_and_assert_signups(
+        api_client, registration2.id, "", [signup4, signup5, signup6, signup7]
+    )
+
+    #  cannot get signups by registration id from an event that is not managed by the user
+    api_client.force_authenticate(user2)
+    response = get_list(api_client, registration.id, "")
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+
+    #  search signups by name
+    get_list_and_assert_signups(api_client, registration2.id, "text=mickey", [signup7])
+
+    #  search signups by membership number
+    get_list_and_assert_signups(
+        api_client, registration2.id, "text=34", [signup6, signup7]
+    )
+    get_list_and_assert_signups(api_client, registration2.id, "text=3456", [signup7])
+
+    #  search signups by extra_info
+    get_list_and_assert_signups(
+        api_client, registration2.id, "text=cd", [signup4, signup5]
+    )
+    get_list_and_assert_signups(api_client, registration2.id, "text=abcd", [signup5])
